chore(*): separating entities and validating strict parity
