+++ AccordProject.Concerto/Introspect/EnumDeclaration.cs
@@ -1,0 +1,38 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using System.Collections.ObjectModel;
+
+namespace AccordProject.Concerto;
+
+/// <summary>
+/// Defines an enumeration of static string values.
+/// Mirrors EnumDeclaration in concerto-core.
+/// </summary>
+public sealed class EnumDeclaration : Declaration
+{
+    private readonly ReadOnlyCollection<string> values;
+
+    internal EnumDeclaration(
+        ModelManager modelManager,
+        string @namespace, string? version, string name,
+        IEnumerable<string> enumValues,
+        IEnumerable<Decorator> decorators)
+        : base(modelManager, @namespace, version, name, decorators)
+    {
+        values = enumValues.ToArray().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<string> GetValues() => values;
+}
